Add class names for notification/status types
I wanted to add user styles that hide certain notifications, so having
class names by type is a big help for that. This commit simply appends
the notification type to a "status-" class or a "notification-" class,
depending on if it's a Status or Notification component. As a result you
get things like "status-mention", "status-favourite", and
"notification-follow".

This could also eventually help with adding a Mastodon-style "filter
your notifications by type" feature.