refactor(themes): use CSS specificity order for themes
The point of this PR is to make it easier to implement scrollbars (#683).

With this PR, the themes move from a body tag-based system (e.g. `body.theme-scarlet`) to a system where they simply declare global CSS and we use CSS specificity order to give us the right theme.

So in this case every theme needs to just declare some CSS variables in
`:root {}` and then the order determines which one wins. We need to
ensure that the order is

```html
<style>/* default styles */</style>
<link>/* theme styles */</style>
<style>/* offline styles */</style>
```

...which is what this PR does.

As a bonus, we can also remove the `fg-loadcss` dependency as I found we
don't really need it.