fix: change dark grayscale to use have a brighter action-button color
The pressed state, such as used for the "Unfollow" button, was effectively
identical to the background, thus making it invisible.

Brighten it up to fix this. The relative differences are kept the same as
before and match the default theme,, e.g. the brightness ranges from
(darkest) Unfollow > Unfollow-hover > Follow > Follow-hover (brightest)

Fixes https://github.com/nolanlawson/pinafore/issues/1880.